requirements.txt
streamlit
numpy
psycopg2-binary==2.8.6